@@ -22,8 +22,8 @@ class BarChartNegative extends StatefulWidget {
 }
 
 class BarChartNegativeState extends SimpleActionState<BarChartNegative> {
-  BarChartController _controller;
-  List<Data> _data = List();
+  late BarChartController _controller;
+  List<Data> _data = [];
 
   @override
   void initState() {
@@ -58,8 +58,8 @@ class BarChartNegativeState extends SimpleActionState<BarChartNegative> {
   }
 
   void _initBarData() {
-    List<BarEntry> values = List();
-    List<Color> colors = List();
+    List<BarEntry> values = [];
+    List<Color> colors = [];
 
     Color green = Color.fromARGB(255, 110, 190, 102);
     Color red = Color.fromARGB(255, 211, 74, 88);
@@ -82,8 +82,8 @@ class BarChartNegativeState extends SimpleActionState<BarChartNegative> {
     set.setColors1(colors);
     set.setValueTextColors(colors);
 
-    _controller.data = BarData(List()..add(set));
-    _controller.data
+    _controller.data = BarData([]..add(set));
+    _controller.data!
       ..setValueTextSize(13)
       ..setValueTypeface(Util.REGULAR)
       ..setValueFormatter(Formatter())
@@ -145,8 +145,8 @@ class A extends ValueFormatter {
   A(this._data) : super();
 
   @override
-  String getFormattedValue1(double value) {
-    return _data[min(max(value.toInt(), 0), _data.length - 1)].xAxisValue;
+  String getFormattedValue1(double? value) {
+    return _data[min(max(value!.toInt(), 0), _data.length - 1)].xAxisValue;
   }
 }
 
@@ -159,14 +159,14 @@ class Data {
 }
 
 class Formatter extends ValueFormatter {
-  NumberFormat _format;
+  late NumberFormat _format;
 
   Formatter() : super() {
     _format = NumberFormat("######.0");
   }
 
   @override
-  String getFormattedValue1(double value) {
+  String getFormattedValue1(double? value) {
     return _format.format(value);
   }
 }
