@@ -22,7 +22,7 @@ class LineChartFilled extends StatefulWidget {
 }
 
 class LineChartFilledState extends SimpleActionState<LineChartFilled> {
-  LineChartController _controller;
+  late LineChartController _controller;
   var random = Random(1);
   int _count = 45;
   double _range = 100.0;
@@ -140,15 +140,15 @@ class LineChartFilledState extends SimpleActionState<LineChartFilled> {
         legendSettingFunction: (legend, controller) {
           legend.enabled = (false);
           var formatter1 = (controller as LineChartController)
-              .data
-              .getDataSetByIndex(0)
+              .data!
+              .getDataSetByIndex(0)!
               .getFillFormatter();
           if (formatter1 is A) {
             formatter1.setPainter(controller);
           }
           var formatter2 = (controller as LineChartController)
-              .data
-              .getDataSetByIndex(1)
+              .data!
+              .getDataSetByIndex(1)!
               .getFillFormatter();
           if (formatter2 is B) {
             formatter2.setPainter(controller);
@@ -170,14 +170,14 @@ class LineChartFilledState extends SimpleActionState<LineChartFilled> {
   }
 
   void _initLineData(int count, double range) {
-    List<Entry> values1 = new List();
+    List<Entry> values1 = [];
 
     for (int i = 0; i < count; i++) {
       double val = (random.nextDouble() * range) + 50;
       values1.add(new Entry(x: i.toDouble(), y: val));
     }
 
-    List<Entry> values2 = new List();
+    List<Entry> values2 = [];
 
     for (int i = 0; i < count; i++) {
       double val = (random.nextDouble() * range) + 450;
@@ -216,8 +216,8 @@ class LineChartFilledState extends SimpleActionState<LineChartFilled> {
     set2.setFillFormatter(B());
 
     // create a data object with the data sets
-    _controller.data = LineData.fromList(List()..add(set1)..add(set2));
-    _controller.data.setDrawValues(false);
+    _controller.data = LineData.fromList([]..add(set1)..add(set2));
+    _controller.data!.setDrawValues(false);
 
     setState(() {});
   }
@@ -226,29 +226,29 @@ class LineChartFilledState extends SimpleActionState<LineChartFilled> {
 }
 
 class A implements IFillFormatter {
-  LineChartController _controller;
+  late LineChartController _controller;
 
   void setPainter(LineChartController controller) {
     _controller = controller;
   }
 
   @override
-  double getFillLinePosition(
-      ILineDataSet dataSet, LineDataProvider dataProvider) {
+  double? getFillLinePosition(
+      ILineDataSet dataSet, LineDataProvider? dataProvider) {
     return _controller?.painter?.axisLeft?.axisMinimum;
   }
 }
 
 class B implements IFillFormatter {
-  LineChartController _controller;
+  late LineChartController _controller;
 
   void setPainter(LineChartController controller) {
     _controller = controller;
   }
 
   @override
-  double getFillLinePosition(
-      ILineDataSet dataSet, LineDataProvider dataProvider) {
+  double? getFillLinePosition(
+      ILineDataSet dataSet, LineDataProvider? dataProvider) {
     return _controller?.painter?.axisLeft?.axisMaximum;
   }
 }
