@@ -19,10 +19,10 @@ class CombinedChart
 class CombinedChartState extends ChartState<CombinedChart> {
   @override
   void updatePainter() {
-    widget.controller.painter.highlightValue6(lastHighlighted, false);
+    widget.controller.painter!.highlightValue6(lastHighlighted, false);
   }
 
-  IDataSet _closestDataSetToTouch;
+  IDataSet? _closestDataSetToTouch;
 
   double _curX = 0.0;
   double _curY = 0.0;
@@ -30,7 +30,7 @@ class CombinedChartState extends ChartState<CombinedChart> {
   bool _isScaleDirectionConfirm = false;
   bool _isYDirection = false;
 
-  Highlight lastHighlighted;
+  Highlight? lastHighlighted;
 
   MPPointF _getTrans(double x, double y) {
     return Utils.local2Chart(widget.controller, x, y, inverted: _inverted());
@@ -49,10 +49,10 @@ class CombinedChartState extends ChartState<CombinedChart> {
 
   bool _inverted() {
     return (_closestDataSetToTouch == null &&
-            widget.controller.painter.isAnyAxisInverted()) ||
+            widget.controller.painter!.isAnyAxisInverted()) ||
         (_closestDataSetToTouch != null &&
-            widget.controller.painter
-                .isInverted(_closestDataSetToTouch.getAxisDependency()));
+            widget.controller.painter!
+                .isInverted(_closestDataSetToTouch!.getAxisDependency()));
   }
 
   @override
@@ -60,23 +60,23 @@ class CombinedChartState extends ChartState<CombinedChart> {
     widget.controller.stopDeceleration();
     _curX = details.localPosition.dx;
     _curY = details.localPosition.dy;
-    _closestDataSetToTouch = widget.controller.painter.getDataSetByTouchPoint(
+    _closestDataSetToTouch = widget.controller.painter!.getDataSetByTouchPoint(
         details.localPosition.dx, details.localPosition.dy);
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPosition.dx,
           details.globalPosition.dy,
           details.localPosition.dx,
           details.localPosition.dy);
-      widget.controller.touchEventListener.onTapDown(point.x, point.y);
+      widget.controller.touchEventListener!.onTapDown(point.x, point.y);
     }
   }
 
   @override
   void onSingleTapUp(TapUpDetails details) {
-    if (widget.controller.painter.highlightPerDragEnabled) {
-      Highlight h = widget.controller.painter.getHighlightByTouchPoint(
+    if (widget.controller.painter!.highlightPerDragEnabled) {
+      Highlight? h = widget.controller.painter!.getHighlightByTouchPoint(
           details.localPosition.dx, details.localPosition.dy);
       lastHighlighted = HighlightUtils.performHighlight(
           widget.controller.painter, h, lastHighlighted);
@@ -86,25 +86,25 @@ class CombinedChartState extends ChartState<CombinedChart> {
     }
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPosition.dx,
           details.globalPosition.dy,
           details.localPosition.dx,
           details.localPosition.dy);
-      widget.controller.touchEventListener.onSingleTapUp(point.x, point.y);
+      widget.controller.touchEventListener!.onSingleTapUp(point.x, point.y);
     }
   }
 
   @override
   void onDoubleTapUp(TapUpDetails details) {
     widget.controller.stopDeceleration();
-    if (widget.controller.painter.doubleTapToZoomEnabled &&
-        widget.controller.painter.getData().getEntryCount() > 0) {
+    if (widget.controller.painter!.doubleTapToZoomEnabled &&
+        widget.controller.painter!.getData()!.getEntryCount() > 0) {
       MPPointF trans =
           _getTrans(details.localPosition.dx, details.localPosition.dy);
-      widget.controller.painter.zoom(
-          widget.controller.painter.scaleXEnabled ? 1.2 : 1,
-          widget.controller.painter.scaleYEnabled ? 1.2 : 1,
+      widget.controller.painter!.zoom(
+          widget.controller.painter!.scaleXEnabled ? 1.2 : 1,
+          widget.controller.painter!.scaleYEnabled ? 1.2 : 1,
           trans.x,
           trans.y);
       setStateIfNotDispose();
@@ -112,12 +112,12 @@ class CombinedChartState extends ChartState<CombinedChart> {
     }
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPosition.dx,
           details.globalPosition.dy,
           details.localPosition.dx,
           details.localPosition.dy);
-      widget.controller.touchEventListener.onDoubleTapUp(point.x, point.y);
+      widget.controller.touchEventListener!.onDoubleTapUp(point.x, point.y);
     }
   }
 
@@ -128,12 +128,12 @@ class CombinedChartState extends ChartState<CombinedChart> {
     _curY = details.localPoint.dy;
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPoint.dx,
           details.globalPoint.dy,
           details.localPoint.dx,
           details.localPoint.dy);
-      widget.controller.touchEventListener.onMoveStart(point.x, point.y);
+      widget.controller.touchEventListener!.onMoveStart(point.x, point.y);
     }
   }
 
@@ -143,8 +143,8 @@ class CombinedChartState extends ChartState<CombinedChart> {
     var dx = details.localPoint.dx - _curX;
     var dy = details.localPoint.dy - _curY;
 
-    if (widget.controller.painter.highlightPerDragEnabled) {
-      final highlighted = widget.controller.painter.getHighlightByTouchPoint(
+    if (widget.controller.painter!.highlightPerDragEnabled) {
+      final highlighted = widget.controller.painter!.getHighlightByTouchPoint(
           details.localPoint.dx, details.localPoint.dy);
       if (highlighted?.equalTo(lastHighlighted) == false) {
         lastHighlighted = HighlightUtils.performHighlight(
@@ -153,34 +153,34 @@ class CombinedChartState extends ChartState<CombinedChart> {
       }
     }
 
-    if (widget.controller.painter.dragYEnabled &&
-        widget.controller.painter.dragXEnabled) {
+    if (widget.controller.painter!.dragYEnabled &&
+        widget.controller.painter!.dragXEnabled) {
       if (_inverted()) {
         dy = -dy;
       }
-      widget.controller.painter.translate(dx, dy);
+      widget.controller.painter!.translate(dx, dy);
       needStateIfNotDispose = true;
     } else {
-      if (widget.controller.painter.dragXEnabled) {
-        widget.controller.painter.translate(dx, 0.0);
+      if (widget.controller.painter!.dragXEnabled) {
+        widget.controller.painter!.translate(dx, 0.0);
         needStateIfNotDispose = true;
-      } else if (widget.controller.painter.dragYEnabled) {
+      } else if (widget.controller.painter!.dragYEnabled) {
         if (_inverted()) {
           dy = -dy;
         }
-        widget.controller.painter.translate(0.0, dy);
+        widget.controller.painter!.translate(0.0, dy);
         needStateIfNotDispose = true;
       }
     }
 
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPoint.dx,
           details.globalPoint.dy,
           details.localPoint.dx,
           details.localPoint.dy);
-      widget.controller.touchEventListener.onMoveUpdate(point.x, point.y);
+      widget.controller.touchEventListener!.onMoveUpdate(point.x, point.y);
     }
 
     if(needStateIfNotDispose){
@@ -199,12 +199,12 @@ class CombinedChartState extends ChartState<CombinedChart> {
       ..computeScroll();
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPoint.dx,
           details.globalPoint.dy,
           details.localPoint.dx,
           details.localPoint.dy);
-      widget.controller.touchEventListener.onMoveEnd(point.x, point.y);
+      widget.controller.touchEventListener!.onMoveEnd(point.x, point.y);
     }
   }
 
@@ -213,12 +213,12 @@ class CombinedChartState extends ChartState<CombinedChart> {
     _scale = -1.0;
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPoint.dx,
           details.globalPoint.dy,
           details.localPoint.dx,
           details.localPoint.dy);
-      widget.controller.touchEventListener.onScaleEnd(point.x, point.y);
+      widget.controller.touchEventListener!.onScaleEnd(point.x, point.y);
     }
   }
 
@@ -230,12 +230,12 @@ class CombinedChartState extends ChartState<CombinedChart> {
     _isScaleDirectionConfirm = false;
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPoint.dx,
           details.globalPoint.dy,
           details.localPoint.dx,
           details.localPoint.dy);
-      widget.controller.touchEventListener.onScaleStart(point.x, point.y);
+      widget.controller.touchEventListener!.onScaleStart(point.x, point.y);
     }
   }
 
@@ -266,28 +266,28 @@ class CombinedChartState extends ChartState<CombinedChart> {
     }
 
     MPPointF trans = _getTrans(_curX, _curY);
-    var h = widget.controller.painter.viewPortHandler;
+    var h = widget.controller.painter!.viewPortHandler;
     scale = Utils.optimizeScale(scale);
     if (pinchZoomEnabled) {
-      bool canZoomMoreX = scale < 1 ? h.canZoomOutMoreX() : h.canZoomInMoreX();
-      bool canZoomMoreY = scale < 1 ? h.canZoomOutMoreY() : h.canZoomInMoreY();
-      widget.controller.painter.zoom(
+      bool canZoomMoreX = scale < 1 ? h!.canZoomOutMoreX() : h!.canZoomInMoreX();
+      bool canZoomMoreY = scale < 1 ? h!.canZoomOutMoreY() : h!.canZoomInMoreY();
+      widget.controller.painter!.zoom(
           canZoomMoreX ? scale : 1, canZoomMoreY ? scale : 1, trans.x, trans.y);
       needStateIfNotDispose = true;
     } else {
       if (_isYDirection) {
-        if (widget.controller.painter.scaleYEnabled) {
+        if (widget.controller.painter!.scaleYEnabled) {
           bool canZoomMoreY =
-              scale < 1 ? h.canZoomOutMoreY() : h.canZoomInMoreY();
-          widget.controller.painter
+              scale < 1 ? h!.canZoomOutMoreY() : h!.canZoomInMoreY();
+          widget.controller.painter!
               .zoom(1, canZoomMoreY ? scale : 1, trans.x, trans.y);
           needStateIfNotDispose = true;
         }
       } else {
-        if (widget.controller.painter.scaleXEnabled) {
+        if (widget.controller.painter!.scaleXEnabled) {
           bool canZoomMoreX =
-              scale < 1 ? h.canZoomOutMoreX() : h.canZoomInMoreX();
-          widget.controller.painter
+              scale < 1 ? h!.canZoomOutMoreX() : h!.canZoomInMoreX();
+          widget.controller.painter!
               .zoom(canZoomMoreX ? scale : 1, 1, trans.x, trans.y);
           needStateIfNotDispose = true;
         }
@@ -296,12 +296,12 @@ class CombinedChartState extends ChartState<CombinedChart> {
 
     if(widget.controller.touchEventListener != null){
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalFocalPoint.dx,
           details.globalFocalPoint.dy,
           details.localFocalPoint.dx,
           details.localFocalPoint.dy);
-      widget.controller.touchEventListener.onScaleUpdate(point.x, point.y);
+      widget.controller.touchEventListener!.onScaleUpdate(point.x, point.y);
     }
 
     if(needStateIfNotDispose) {
@@ -320,36 +320,36 @@ class CombinedChartState extends ChartState<CombinedChart> {
   void onDragStart(LongPressStartDetails details) {
     if (widget.controller.touchEventListener != null) {
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPosition.dx,
           details.globalPosition.dy,
           details.localPosition.dx,
           details.localPosition.dy);
-      widget.controller.touchEventListener.onDragStart(point.x, point.y);
+      widget.controller.touchEventListener!.onDragStart(point.x, point.y);
     }
   }
 
   void onDragUpdate(LongPressMoveUpdateDetails details) {
     if (widget.controller.touchEventListener != null) {
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPosition.dx,
           details.globalPosition.dy,
           details.localPosition.dx,
           details.localPosition.dy);
-      widget.controller.touchEventListener.onDragUpdate(point.x, point.y);
+      widget.controller.touchEventListener!.onDragUpdate(point.x, point.y);
     }
   }
 
   void onDragEnd(LongPressEndDetails details) {
     if (widget.controller.touchEventListener != null) {
       var point = _getTouchValue(
-          widget.controller.touchEventListener.valueType(),
+          widget.controller.touchEventListener!.valueType(),
           details.globalPosition.dx,
           details.globalPosition.dy,
           details.localPosition.dx,
           details.localPosition.dy);
-      widget.controller.touchEventListener.onDragEnd(point.x, point.y);
+      widget.controller.touchEventListener!.onDragEnd(point.x, point.y);
     }
   }
 }
