@@ -5,6 +5,7 @@ import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
 import 'package:mp_chart/mp/chart/bar_chart.dart';
 import 'package:mp_chart/mp/controller/bar_chart_controller.dart';
+import 'package:mp_chart/mp/controller/bar_line_scatter_candle_bubble_controller.dart';
 import 'package:mp_chart/mp/core/color/gradient_color.dart';
 import 'package:mp_chart/mp/core/common_interfaces.dart';
 import 'package:mp_chart/mp/core/data/bar_data.dart';
@@ -180,7 +181,7 @@ class BarChartBasicState extends BarActionState<BarChartBasic>
             ..drawGridLines = false
             ..setGranularity(1.0)
             ..setLabelCount1(7)
-            ..setValueFormatter(DayAxisValueFormatter(controller));
+            ..setValueFormatter(DayAxisValueFormatter(controller as BarLineScatterCandleBubbleController));
         },
         selectionListener: this,
         drawBarShadow: false,
@@ -199,7 +200,7 @@ class BarChartBasicState extends BarActionState<BarChartBasic>
   void _initData(int count, double range, ui.Image img) {
     double start = 1;
 
-    List<BarEntry> values = List();
+    List<BarEntry> values = [];
 
     for (int i = start.toInt(); i < start + count; i++) {
       double val = (random.nextDouble() * (range + 1));
@@ -230,7 +231,7 @@ class BarChartBasicState extends BarActionState<BarChartBasic>
     Color endColor4 = ColorUtils.HOLO_RED_DARK;
     Color endColor5 = ColorUtils.HOLO_ORANGE_DARK;
 
-    List<GradientColor> gradientColors = List();
+    List<GradientColor> gradientColors = [];
     gradientColors.add(GradientColor(startColor1, endColor1));
     gradientColors.add(GradientColor(startColor2, endColor2));
     gradientColors.add(GradientColor(startColor3, endColor3));
@@ -239,11 +240,11 @@ class BarChartBasicState extends BarActionState<BarChartBasic>
 
     set1.setGradientColors(gradientColors);
 
-    List<IBarDataSet> dataSets = List();
+    List<IBarDataSet> dataSets = [];
     dataSets.add(set1);
 
     controller.data = BarData(dataSets);
-    controller.data
+    controller.data!
       ..setValueTextSize(10)
       ..setValueTypeface(Util.LIGHT)
       ..barWidth = 0.9;
@@ -280,7 +281,7 @@ class BarChartBasicState extends BarActionState<BarChartBasic>
 
 class MyChartTransListener with ChartTransListener {
   @override
-  void scale(double scaleX, double scaleY, double x, double y) {
+  void scale(double scaleX, double scaleY, double? x, double? y) {
     print("scale scaleX: $scaleX, scaleY: $scaleY, x: $x, y: $y");
   }
 
@@ -292,47 +293,47 @@ class MyChartTransListener with ChartTransListener {
 
 class MyTouchEventListener with OnTouchEventListener {
   @override
-  void onDoubleTapUp(double x, double y) {
+  void onDoubleTapUp(double? x, double? y) {
     print("onDoubleTapUp x: $x, y: $y");
   }
 
   @override
-  void onMoveEnd(double x, double y) {
+  void onMoveEnd(double? x, double? y) {
     print("onMoveEnd x: $x, y: $y");
   }
 
   @override
-  void onMoveStart(double x, double y) {
+  void onMoveStart(double? x, double? y) {
     print("onMoveStart x: $x, y: $y");
   }
 
   @override
-  void onMoveUpdate(double x, double y) {
+  void onMoveUpdate(double? x, double? y) {
     print("onMoveUpdate x: $x, y: $y");
   }
 
   @override
-  void onScaleEnd(double x, double y) {
+  void onScaleEnd(double? x, double? y) {
     print("onScaleEnd x: $x, y: $y");
   }
 
   @override
-  void onScaleStart(double x, double y) {
+  void onScaleStart(double? x, double? y) {
     print("onScaleStart x: $x, y: $y");
   }
 
   @override
-  void onScaleUpdate(double x, double y) {
+  void onScaleUpdate(double? x, double? y) {
     print("onScaleUpdate x: $x, y: $y");
   }
 
   @override
-  void onSingleTapUp(double x, double y) {
+  void onSingleTapUp(double? x, double? y) {
     print("onSingleTapUp x: $x, y: $y");
   }
 
   @override
-  void onTapDown(double x, double y) {
+  void onTapDown(double? x, double? y) {
     print("onTapDown x: $x, y: $y");
   }
 
@@ -342,17 +343,17 @@ class MyTouchEventListener with OnTouchEventListener {
   }
 
   @override
-  void onDragEnd(double x, double y) {
+  void onDragEnd(double? x, double? y) {
     print("onDragEnd x: $x, y: $y");
   }
 
   @override
-  void onDragStart(double x, double y) {
+  void onDragStart(double? x, double? y) {
     print("onDragStart x: $x, y: $y");
   }
 
   @override
-  void onDragUpdate(double x, double y) {
+  void onDragUpdate(double? x, double? y) {
     print("onDragUpdate x: $x, y: $y");
   }
 }
