@@ -123,7 +123,7 @@ class BarChartBasic2State extends BarActionState<BarChartBasic2> {
 
   void _initBarData(int count, double range) async {
     var img = await ImageLoader.loadImage('assets/img/star.png');
-    List<BarEntry> values = List();
+    List<BarEntry> values = [];
 
     for (int i = 0; i < count; i++) {
       double multi = (range + 1);
@@ -137,11 +137,11 @@ class BarChartBasic2State extends BarActionState<BarChartBasic2> {
     set1.setColors1(ColorUtils.VORDIPLOM_COLORS);
     set1.setDrawValues(false);
 
-    List<IBarDataSet> dataSets = List();
+    List<IBarDataSet> dataSets = [];
     dataSets.add(set1);
 
     controller.data = BarData(dataSets);
-    controller.data
+    controller.data!
       ..setValueTextSize(10)
       ..barWidth = 0.9;
 
@@ -176,7 +176,7 @@ class BarChartBasic2State extends BarActionState<BarChartBasic2> {
 
   Widget _initBarChart() {
     var barChart = BarChart(controller);
-    controller.animator
+    controller.animator!
       ..reset()
       ..animateY1(1500);
     return barChart;
