@@ -188,13 +188,13 @@ class BarChartMultipleState extends BarActionState<BarChartMultiple>
         description: desc);
   }
 
-  int groupCount;
-  int startYear;
-  int endYear;
+  int groupCount = 0;
+  int startYear = 0;
+  int endYear = 0;
   bool isDataInitial = false;
 
   void _initBarData(int count, double range) async {
-    List<ui.Image> imgs = List(3);
+    List<ui.Image?> imgs = List.filled(3,null);
     imgs[0] = await ImageLoader.loadImage('assets/img/star.png');
     imgs[1] = await ImageLoader.loadImage('assets/img/add.png');
     imgs[2] = await ImageLoader.loadImage('assets/img/close.png');
@@ -202,10 +202,10 @@ class BarChartMultipleState extends BarActionState<BarChartMultiple>
     startYear = 1980;
     endYear = startYear + groupCount;
 
-    List<BarEntry> values1 = List();
-    List<BarEntry> values2 = List();
-    List<BarEntry> values3 = List();
-    List<BarEntry> values4 = List();
+    List<BarEntry> values1 = [];
+    List<BarEntry> values2 = [];
+    List<BarEntry> values3 = [];
+    List<BarEntry> values4 = [];
 
     double randomMultiplier = range * 100000;
 
@@ -239,8 +239,8 @@ class BarChartMultipleState extends BarActionState<BarChartMultiple>
     set4.setColor1(Color.fromARGB(255, 255, 102, 0));
 
     controller.data =
-        BarData(List()..add(set1)..add(set2)..add(set3)..add(set4));
-    controller.data
+        BarData([]..add(set1)..add(set2)..add(set3)..add(set4));
+    controller.data!
       ..setValueFormatter(LargeValueFormatter())
       ..setValueTypeface(Util.LIGHT)
       // specify the width each bar should have
@@ -260,7 +260,7 @@ class BarChartMultipleState extends BarActionState<BarChartMultiple>
 
 class A extends ValueFormatter {
   @override
-  String getFormattedValue1(double value) {
-    return value.toInt().toString();
+  String getFormattedValue1(double? value) {
+    return value!.toInt().toString();
   }
 }
