@@ -198,7 +198,7 @@ class LineChartBasicState extends LineActionState<LineChartBasic> {
 
   void _initLineData(int count, double range) async {
     var img = await ImageLoader.loadImage('assets/img/star.png');
-    List<Entry> values = List();
+    List<Entry> values = [];
 
     for (int i = 0; i < count; i++) {
       double val = (random.nextDouble() * range) - 30;
@@ -245,7 +245,7 @@ class LineChartBasicState extends LineActionState<LineChartBasic> {
     // set color of filled area
     set1.setGradientColor(ColorUtils.BLUE, ColorUtils.RED);
 
-    List<ILineDataSet> dataSets = List();
+    List<ILineDataSet> dataSets = [];
     dataSets.add(set1); // add the data sets
 
     // create a data object with the data sets
@@ -256,7 +256,7 @@ class LineChartBasicState extends LineActionState<LineChartBasic> {
 
   Widget _initLineChart() {
     var lineChart = LineChart(controller);
-    controller.animator
+    controller.animator!
       ..reset()
       ..animateX1(1500);
     return lineChart;
