@@ -19,12 +19,12 @@ class LineChartColorful extends StatefulWidget {
 }
 
 class LineChartColorfulState extends SimpleActionState<LineChartColorful> {
-  List<LineChartController> _controllers = List(4);
+  List<LineChartController?> _controllers = List.filled(4,null);
   var random = Random(1);
   int _count = 36;
   double _range = 100.0;
 
-  List<Color> _colors = List()
+  List<Color> _colors = []
     ..add(Color.fromARGB(255, 137, 230, 81))
     ..add(Color.fromARGB(255, 240, 240, 30))
     ..add(Color.fromARGB(255, 89, 199, 250))
@@ -54,19 +54,19 @@ class LineChartColorfulState extends SimpleActionState<LineChartColorful> {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               Expanded(
-                child: getLineChart(_controllers[0]),
+                child: getLineChart(_controllers[0]!),
                 flex: 1,
               ),
               Expanded(
-                child: getLineChart(_controllers[1]),
+                child: getLineChart(_controllers[1]!),
                 flex: 1,
               ),
               Expanded(
-                child: getLineChart(_controllers[2]),
+                child: getLineChart(_controllers[2]!),
                 flex: 1,
               ),
               Expanded(
-                child: getLineChart(_controllers[3]),
+                child: getLineChart(_controllers[3]!),
                 flex: 1,
               ),
             ],
@@ -84,15 +84,15 @@ class LineChartColorfulState extends SimpleActionState<LineChartColorful> {
 
   void _initLineData(int count, double range) {
     for (int i = 0; i < _controllers.length; i++) {
-      _controllers[i].data = _getData(36, 100);
-      _controllers[i].data.setValueTypeface(Util.BOLD);
-      (_controllers[i].data.getDataSetByIndex(0) as LineDataSet)
+      _controllers[i]!.data = _getData(36, 100);
+      _controllers[i]!.data!.setValueTypeface(Util.BOLD);
+      (_controllers[i]!.data!.getDataSetByIndex(0) as LineDataSet)
           .setCircleHoleColor(_colors[i % _colors.length]);
     }
   }
 
   LineData _getData(int count, double range) {
-    List<Entry> values = List();
+    List<Entry> values = [];
 
     for (int i = 0; i < count; i++) {
       double val = (random.nextDouble() * range) + 3;
@@ -113,7 +113,7 @@ class LineChartColorfulState extends SimpleActionState<LineChartColorful> {
     set1.setDrawValues(false);
 
     // create a data object with the data sets
-    return LineData.fromList(List()..add(set1));
+    return LineData.fromList([]..add(set1));
   }
 
   LineChartController _setupChartController(Color color) {
@@ -148,7 +148,7 @@ class LineChartColorfulState extends SimpleActionState<LineChartColorful> {
 
   Widget getLineChart(LineChartController controller) {
     var lineChart = LineChart(controller);
-    controller.animator
+    controller.animator!
       ..reset()
       ..animateX1(2500);
     return lineChart;
