@@ -135,9 +135,9 @@ class LineChartDualAxisState extends LineActionState<LineChartDualAxis>
   @override
   void onValueSelected(Entry e, Highlight h) {
     controller?.centerViewToAnimated(
-        e.x,
-        e.y,
-        controller.data.getDataSetByIndex(h.dataSetIndex).getAxisDependency(),
+        e.x!,
+        e.y!,
+        controller.data!.getDataSetByIndex(h.dataSetIndex)!.getAxisDependency(),
         500);
   }
 
@@ -197,25 +197,25 @@ class LineChartDualAxisState extends LineActionState<LineChartDualAxis>
   }
 
   void _initLineData(int count, double range) async {
-    List<ui.Image> imgs = List(3);
+    List<ui.Image?> imgs = List.filled(3,null);
     imgs[0] = await ImageLoader.loadImage('assets/img/star.png');
     imgs[1] = await ImageLoader.loadImage('assets/img/add.png');
     imgs[2] = await ImageLoader.loadImage('assets/img/close.png');
-    List<Entry> values1 = List();
+    List<Entry> values1 = [];
 
     for (int i = 0; i < count; i++) {
       double val = (random.nextDouble() * (range / 2.0)) + 50;
       values1.add(Entry(x: i.toDouble(), y: val, icon: imgs[0]));
     }
 
-    List<Entry> values2 = List();
+    List<Entry> values2 = [];
 
     for (int i = 0; i < count; i++) {
       double val = (random.nextDouble() * range) + 450;
       values2.add(new Entry(x: i.toDouble(), y: val, icon: imgs[1]));
     }
 
-    List<Entry> values3 = List();
+    List<Entry> values3 = [];
 
     for (int i = 0; i < count; i++) {
       double val = (random.nextDouble() * range) + 500;
@@ -268,8 +268,8 @@ class LineChartDualAxisState extends LineActionState<LineChartDualAxis>
 
     // create a data object with the data sets
     controller.data =
-        LineData.fromList(List()..add(set1)..add(set2)..add(set3));
-    controller.data
+        LineData.fromList([]..add(set1)..add(set2)..add(set3));
+    controller.data!
       ..setValueTextColor(ColorUtils.WHITE)
       ..setValueTextSize(9);
 
@@ -278,7 +278,7 @@ class LineChartDualAxisState extends LineActionState<LineChartDualAxis>
 
   Widget _initLineChart() {
     var lineChart = LineChart(controller);
-    controller.animator
+    controller.animator!
       ..reset()
       ..animateX1(1500);
     return lineChart;
