@@ -31,7 +31,7 @@ class LineChartMultipleState extends LineActionState<LineChartMultiple>
   int _count = 20;
   double _range = 100.0;
 
-  List<Color> colors = List()
+  List<Color> colors = []
     ..add(ColorUtils.VORDIPLOM_COLORS[0])
     ..add(ColorUtils.VORDIPLOM_COLORS[1])
     ..add(ColorUtils.VORDIPLOM_COLORS[2]);
@@ -198,14 +198,14 @@ class LineChartMultipleState extends LineActionState<LineChartMultiple>
   }
 
   void _initLineData(int count, double range) async {
-    List<ui.Image> imgs = List(3);
+    List<ui.Image?> imgs = List.filled(3,null);
     imgs[0] = await ImageLoader.loadImage('assets/img/star.png');
     imgs[1] = await ImageLoader.loadImage('assets/img/add.png');
     imgs[2] = await ImageLoader.loadImage('assets/img/close.png');
-    List<ILineDataSet> dataSets = List();
+    List<ILineDataSet> dataSets = [];
 
     for (int z = 0; z < 3; z++) {
-      List<Entry> values = List();
+      List<Entry> values = [];
 
       for (int i = 0; i < count; i++) {
         double val = (random.nextDouble() * range) + 3;
