@@ -168,13 +168,13 @@ class OtherChartBubbleState extends BubbleActionState<OtherChartBubble>
   }
 
   void _initBubbleData(int count, double range) async {
-    List<ui.Image> imgs = List(3);
+    List<ui.Image?> imgs = List.filled(3,null);
     imgs[0] = await ImageLoader.loadImage('assets/img/star.png');
     imgs[1] = await ImageLoader.loadImage('assets/img/add.png');
     imgs[2] = await ImageLoader.loadImage('assets/img/close.png');
-    List<BubbleEntry> values1 = List();
-    List<BubbleEntry> values2 = List();
-    List<BubbleEntry> values3 = List();
+    List<BubbleEntry> values1 = [];
+    List<BubbleEntry> values2 = [];
+    List<BubbleEntry> values3 = [];
 
     for (int i = 0; i < count; i++) {
       values1.add(BubbleEntry(
@@ -210,14 +210,14 @@ class OtherChartBubbleState extends BubbleActionState<OtherChartBubble>
     set3.setColor3(ColorUtils.COLORFUL_COLORS[2], 130);
     set3.setDrawValues(true);
 
-    List<IBubbleDataSet> dataSets = List();
+    List<IBubbleDataSet> dataSets = [];
     dataSets.add(set1); // add the data sets
     dataSets.add(set2);
     dataSets.add(set3);
 
     // create a data object with the data sets
     controller.data = BubbleData.fromList(dataSets);
-    controller.data
+    controller.data!
       ..setDrawValues(false)
       ..setValueTypeface(Util.LIGHT)
       ..setValueTextSize(8)
