@@ -38,7 +38,7 @@ class ScrollingChartMultiple extends StatefulWidget {
 
 class ScrollingChartMultipleState
     extends SimpleActionState<ScrollingChartMultiple> {
-  List<Controller> _controllers = List();
+  List<Controller> _controllers = [];
   var random = Random(1);
   bool _isParentMove = true;
   double _curX = 0.0;
@@ -106,11 +106,13 @@ class ScrollingChartMultipleState
   Widget _renderItem(int index) {
     Chart chart;
     if (_controllers[index] is LineChartController) {
-      chart = _getLineChart(_controllers[index]);
+      chart = _getLineChart(_controllers[index] as LineChartController);
     } else if (_controllers[index] is BarChartController) {
-      chart = _getBarChart(_controllers[index]);
+      chart = _getBarChart(_controllers[index] as BarChartController);
     } else if (_controllers[index] is PieChartController) {
-      chart = _getPieChart(_controllers[index]);
+      chart = _getPieChart(_controllers[index] as PieChartController );
+    }else{
+      chart = _getLineChart(_controllers[index] as LineChartController);
     }
 
     return Container(height: 200, child: chart);
@@ -210,7 +212,7 @@ class ScrollingChartMultipleState
   }
 
   LineData _generateDataLine(int cnt) {
-    List<Entry> values1 = List();
+    List<Entry> values1 = [];
 
     for (int i = 0; i < 12; i++) {
       values1.add(Entry(x: i.toDouble(), y: (random.nextDouble() * 65) + 40));
@@ -222,10 +224,10 @@ class ScrollingChartMultipleState
     d1.setHighLightColor(Color.fromARGB(255, 244, 117, 117));
     d1.setDrawValues(false);
 
-    List<Entry> values2 = List();
+    List<Entry> values2 = [];
 
     for (int i = 0; i < 12; i++) {
-      values2.add(Entry(x: i.toDouble(), y: values1[i].y - 30));
+      values2.add(Entry(x: i.toDouble(), y: values1[i].y! - 30));
     }
 
     LineDataSet d2 = LineDataSet(values2, "New DataSet $cnt, (2)");
@@ -236,7 +238,7 @@ class ScrollingChartMultipleState
     d2.setCircleColor(ColorUtils.VORDIPLOM_COLORS[0]);
     d2.setDrawValues(false);
 
-    List<ILineDataSet> sets = List();
+    List<ILineDataSet> sets = [];
     sets.add(d1);
     sets.add(d2);
 
@@ -244,7 +246,7 @@ class ScrollingChartMultipleState
   }
 
   BarData _generateDataBar(int cnt) {
-    List<BarEntry> entries = List();
+    List<BarEntry> entries = [];
 
     for (int i = 0; i < 12; i++) {
       entries
@@ -255,13 +257,13 @@ class ScrollingChartMultipleState
     d.setColors1(ColorUtils.VORDIPLOM_COLORS);
     d.setHighLightAlpha(255);
 
-    BarData cd = BarData(List()..add(d));
+    BarData cd = BarData([]..add(d));
     cd.barWidth = (0.9);
     return cd;
   }
 
   PieData _generateDataPie() {
-    List<PieEntry> entries = List();
+    List<PieEntry> entries = [];
 
     for (int i = 0; i < 4; i++) {
       entries.add(PieEntry(
@@ -279,7 +281,7 @@ class ScrollingChartMultipleState
 
   LineChart _getLineChart(LineChartController controller) {
     var lineChart = LineChart(controller);
-    controller.animator
+    controller.animator!
       ..reset()
       ..animateX1(750);
     return lineChart;
@@ -287,19 +289,19 @@ class ScrollingChartMultipleState
 
   BarChart _getBarChart(BarChartController controller) {
     var barChart = BarChart(controller);
-    controller.animator
+    controller.animator!
       ..reset()
       ..animateY1(700);
     return barChart;
   }
 
   PieChart _getPieChart(PieChartController controller) {
-    controller.data
+    controller.data!
       ..setValueFormatter(PercentFormatter())
       ..setValueTextSize(11)
       ..setValueTextColor(ColorUtils.WHITE);
     var pieChart = PieChart(controller);
-    controller.animator
+    controller.animator!
       ..reset()
       ..animateY1(900);
     return pieChart;
