@@ -10,7 +10,7 @@ void main() {
     expect(calculator.addOne(2), 3);
     expect(calculator.addOne(-7), -6);
     expect(calculator.addOne(0), 1);
-    expect(() => calculator.addOne(null), throwsNoSuchMethodError);
+    expect(() => calculator.addOne(1), throwsNoSuchMethodError);
   });
 
   /**
